+++ src/app/pages/linea/linea.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {LineaService} from '../../services/services.index';
 import {Router} from '@angular/router';
 import * as printJS from 'print-js';
+import swal from 'sweetalert';
 @Component({
   selector: 'app-linea',
   templateUrl: './linea.component.html',
@@ -12,6 +13,7 @@ export class LineaComponent implements OnInit {
  public someJSONdata:any=[];
  public date= new Date();
   public fecha ;
+  public page:number=1;
   constructor(private lineaService:LineaService, private router: Router) { 
     this.fecha=(this.date.getDate() + "/" + (this.date.getMonth() +1) + "/" + this.date.getFullYear());
   }
@@ -32,14 +34,36 @@ export class LineaComponent implements OnInit {
   }
 
   deleteLinea(id:number | string){
-    this.lineaService.deleteLinea(id)
+    swal({
+      title:'Eliminar',
+      text: '¿Seguro de eliminar la linea?',
+      icon:'warning',
+      closeOnClickOutside: false,
+      closeOnEsc: false,
+      buttons: {
+        cancelar: {text:'Cancelar',className:'sweet-warning'},
+        confirmar: {text:'Confirmar',className:'sweet-success'},
+      },
+    })
+    .then((value) => {
+      if(value==='confirmar'){
+        this.lineaService.deleteLinea(id)
           .subscribe(
             res=>{
-              
+              swal('Perfecto','La linea fue borrada con exito','success');
               this.getList();
             },
             err=>console.error(err)
           );
+      }
+      if(value==='cancelar'){
+        swal.close();
+      }
+      
+     
+    });
+
+    
   }
   generatePDF(){
