+++ src/app/pages/proveedor/proveedor.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {ProveedorService} from '../../services/services.index';
 import {Router} from '@angular/router';
 import * as printJS from 'print-js';
+import swal from 'sweetalert';
 @Component({
   selector: 'app-proveedor',
   templateUrl: './proveedor.component.html',
@@ -12,6 +13,7 @@ export class ProveedorComponent implements OnInit {
   public someJSONdata:any=[];
   public date= new Date();
   public fecha ;
+  public page: number=1;
   constructor(private proveedorServices: ProveedorService, private router: Router) { 
     this.fecha=(this.date.getDate() + "/" + (this.date.getMonth() +1) + "/" + this.date.getFullYear());
   }
@@ -31,14 +33,35 @@ export class ProveedorComponent implements OnInit {
   }
 
   deleteProv(id:number | string){
-    this.proveedorServices.deleteProv(id)
+    swal({
+      title:'Eliminar',
+      text: '¿Seguro de eliminar el proveedor?',
+      icon:'warning',
+      closeOnClickOutside: false,
+      closeOnEsc: false,
+      buttons: {
+        cancelar: {text:'Cancelar',className:'sweet-warning'},
+        confirmar: {text:'Confirmar',className:'sweet-success'},
+      },
+    })
+      .then((value) => {
+        if(value==='confirmar'){
+          this.proveedorServices.deleteProv(id)
           .subscribe(
             res=>{
-              console.log(res);
+              swal('Perfecto','El proveedor fue eliminado con exito','success');
               this.getList();
             },
             err=>console.error(err)
           );
+        }
+        if(value==='cancelar'){
+          swal.close();
+        }
+        
+      });
+
+    
   }
   generatePDF(){
